+++ Source/RenderPasses/CreateRenderShadowMapCommandsPass.cpp
@@ -55,7 +55,7 @@ CreateRenderShadowMapCommandsPass::CreateRenderShadowMapCommandsPass(InitParams*
 	}
 	
 	D3D12_ROOT_PARAMETER rootParams[kNumRootParams];
-	rootParams[kSRVRootParam] = RootDescriptorTableParameter(srvDescriptorRanges.size(), &srvDescriptorRanges[0], D3D12_SHADER_VISIBILITY_ALL);
+	rootParams[kSRVRootParam] = RootDescriptorTableParameter((UINT)srvDescriptorRanges.size(), &srvDescriptorRanges[0], D3D12_SHADER_VISIBILITY_ALL);
 
 	RootSignatureDesc rootSignatureDesc(kNumRootParams, rootParams);
 	m_pRootSignature = new RootSignature(pRenderEnv->m_pDevice, &rootSignatureDesc, L"CreateRenderShadowMapCommandsPass::m_pRootSignature");
